+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Unreleased
 
+### Changed
+
+- **[BREAKING]** Introduces an `OnException` primitive for `Mask` that replaces `Bracket` and `BracketOnError`
+  (note that this only affects the effect datatype; there is still `bracket` and `bracketOnError` *functions* with the same semantics)
+
 ### Added
 
 - `freshEnumToState` for `Fresh`
+- `onException` and `bracketOnError_` for `Mask`
 
 ## 0.2.1.0 (2022-02-13)